refactor: move initial scene logic to new hook
